issue #994 remove MessageViewController
